Fix script/style enqueue path when using BP dev. repo. with release build of WP, and SCRIPT_DEBUG=false.
If you have a WordPress release build (that is, not the dev SVN with its src folder) with a BuddyPress dev build (with its src folder), and set SCRIPT_DEBUG=false, BuddyPress will try to load the minified versions of its CSS and JS assets. As these minified assets only exist in our release builds, this causes 404 errors in this particular set of circumstances.

Fixes #6920

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10613 cdf35c40-ae34-48e0-9cc9-0c9da1808c22