URI: Support members shortlink redirection.
This commit redirects `/members/me/*` to `/members/LOGGEDIN_USER_SLUG/*`.
If root profiles is enabled, then this redirects `/me/*` to
`/LOGGEDIN_USER_SLUG/*`.  If a user is logged out, `/me/` redirects to the
login page and when properly authenticated, will redirect back to the
logged-in user's page.

This will allow devs to create dynamic links to a user's profile, which is
useful for all sorts of things -- nav menus, support links, etc.

Props r-a-y, dcavins.

Fixes #6325.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10791 cdf35c40-ae34-48e0-9cc9-0c9da1808c22