Notifications: Ensure moderators are able to manage a user's notifications.
Previously, due to hardcoded logged-in user checks, it was not possible
for moderators to manage a displayed user's notifications.

This commit adds a user ID parameter to all appropriate notification
template tag functions and sets some of those functions to default to the
displayed user ID where necessary.  This will allow moderators to manage
a displayed user's notifications as intended.

Props r-a-y, VibeThemes.

Fixes #7023.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10816 cdf35c40-ae34-48e0-9cc9-0c9da1808c22