Make sure the password is set to what user chose when signing up with a blog
When signing up, users choose their password and we are hashing them into the meta field of the signups table.
When users activate their accounts, we directly run a query to replace the password automatically generated by WordPress to what the users chose. For the "signup with a blog" case, it appears this is not taking in account because the created users is cached and the `xprofile_sync_wp_profile()` function is reseting the password to the cached version.
To avoid this, as soon as we run the direct query, we are cleaning the cached object of the created user.

Fixes #7173 (2.6 branch)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.6@10939 cdf35c40-ae34-48e0-9cc9-0c9da1808c22