WP Users List: Add member type bulk change controls and filters.
Add a dropdown select to change the member type of many members in
bulk. Also add a new column, “Member Type,” showing the member’s
assigned member type as a link that leads to a filtered view of the WP
Users table, much like a tag link on the posts table works.

Props slaFFik, dcavins.

Fixes #6060.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11111 cdf35c40-ae34-48e0-9cc9-0c9da1808c22