xprofile: fix regression being unable to select year in date field type.
Pre-2.7 date fields do not have all of the new field meta properties set, and the requested metas were returned as an empty string, not a boolean false. The type strictness of the check changed in this commit caused the value(s) of the year field to be set to an empty array, instead of picking up the range of dates from the defaults.

This was probably caused by some misleading PHPDoc for `bp_xprofile_get_meta`, which we're going to look into seperately.

Fixes #7300

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11210 cdf35c40-ae34-48e0-9cc9-0c9da1808c22