Administration: filter number of members in a list for group.
When editing a group in wp-admin area, we have 2 separate values that control the per page number of members (which should always be in sync). This commit adds a a new filter `bp_groups_admin_members_type_per_page` that gives ability to filter that number using only one `add_filter()`.

See #7405.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11346 cdf35c40-ae34-48e0-9cc9-0c9da1808c22