Build: allow tests on Travis-CI's PHP 7.1 to fail temporarily.
There is an upstream bug in WordPress causing fatal errors for one of our tests in PHP 7.1.
It has not been fixed yet, and as I suspect the test itself is better suited for WordPress than BuddyPress (BuddyPress ought to assume that basic shortcode parsing is functional), I think this is a worthwhile tradeoff for now.

See https://core.trac.wordpress.org/ticket/39277 and #7392

Props netweb

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11391 cdf35c40-ae34-48e0-9cc9-0c9da1808c22