Show displayed user's messages on messages screens.
When a site administrator visits another user’s messages, she is shown
her own messages because the display logic for messages refers to the
logged-in user rather than the displayed user. This changeset updates
the logic to allow the site admin to view and interact with another
user’s messages.

Fixes #7096.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11578 cdf35c40-ae34-48e0-9cc9-0c9da1808c22