Attachments: Fix issues with uploading a file containing accented characters.
Previously, when uploading a new avatar or cover image with accented
characters, the upload would fail.  This is due to an upstream WordPress
bug since WP does not handle sanitizing these types of file names properly.
(See #WP22363, #WP24661, #WP15955 for just a few of the related problems.)

To fix this in BuddyPress, we are putting in a workaround to convert file
names with UTF-8 characters to their ASCII equivalent.  First, we try to
use PHP's `intl` extension if available.  If not, we use WP's
`remove_accents()` function, and lastly if the filename still contains
UTF-8 characters, we also try to use `iconv()` if available.

This is considered a workaround until WordPress fixes this in core.

Fixes #7484.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11602 cdf35c40-ae34-48e0-9cc9-0c9da1808c22