bp-legacy: Fix the display of hidden groups in group loops in some instances.
Themes or plugins that use the `.hidden` CSS declaration to hide content
can conflict with hidden groups in group loops.  Themes that utilize the
Bootstrap framework are particularly prone to this problem.

The issue is due to us using the group status as a CSS class and `'hidden'`
is a valid BuddyPress group status option.

To workaround this, we check the visibility of a hidden group in the group
loop with JS.  If the group is hidden via CSS, we force the display of the
group with the `!important` rule.

In the future, we should probably deprecate the usage of the `hidden` CSS
class in group loops for something else.

Props r-a-y, MikeGillihan, hnla, henry.wright.

Fixes #7443.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11613 cdf35c40-ae34-48e0-9cc9-0c9da1808c22