Activation: Ensure we use the XProfile field's default visibility level during activation.
Previously, if the visibility option for a XProfile field was removed from
the registration template, the visibility level for the user would always
fallback to 'public' instead of what the actual XProfile field's default
visibility level was.

Commit includes a unit test and rectifies the problem.

Props uscore713.

Fixes #7553 (trunk)

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11630 cdf35c40-ae34-48e0-9cc9-0c9da1808c22