Emails: Only add the 'List-Unsubscribe' header if we have a valid unsubscribe link.
Previously, we were adding a blank 'List-Unsubscribe' header for activation
and change of email address emails.  This caused delivery issues for those
using Outlook.com as their email address.  It appears that Outlook.com
requires a valid 'List-Unsubscribe' header in order for the email to be
received.

This commit fixes this by only adding the 'List-Unsubscribe' header if we
have a valid email unsubscribe link.

Props jcrr.

See #7697 (2.9-branch)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.9@11862 cdf35c40-ae34-48e0-9cc9-0c9da1808c22