Core: Introduce hook, `'bp_late_include'`.
This hook is meant to be used to load conditional files on certain pages
so we are not always loading code unnecessarily.

This fires on the `'bp_setup_canonical_stack'` at priority 20 when we
have ensured that the canonical stack is set up and before we register
navigation items that do checks for screen functions.

We'll be using this hook across all components to offshore code until it
is required.

See #7218.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11884 cdf35c40-ae34-48e0-9cc9-0c9da1808c22