bp-legacy: Disable "Send Reply" button onclick when replying to a private message thread via AJAX.
Previously, it was possible for a user to click on the "Send Reply" button
multiple times, causing the same reply to be duplicated several times over.

This commit addreses this by disabling the "Send Reply" button when it is
clicked and re-enables it after the AJAX request is completed.

Props jdgrimes.

Fixes #7718.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11911 cdf35c40-ae34-48e0-9cc9-0c9da1808c22