Groups: Wrap `groups_record_activity()` calls with `bp_is_active( 'activity' )`
In r11924, the `bp-groups-activity.php` file is loaded only if the Activity
component is enabled.

However, the `groups_record_activity()` function is located in this file.
Because this function now relies on the Activity component, if the Activity
component is disabled, a fatal error can occur during certain group
actions.  This commit addresses the problem by checking for the Activity
component before calling on the function.

Anti-props r-a-y.

See #7870 (3.x branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12125 cdf35c40-ae34-48e0-9cc9-0c9da1808c22