Blogs: rename meta table key for compatibilty with #WP37923
The WordPress change introduces a blogmeta table in WordPress 5.0.
It uses `$wpdb->blogmeta` to access the table name, which is what BuddyPress has been using.

This'll be a breaking change for any manual SQL queries as soon as WP 5.0 is released.
When BuddyPress requires 5.0, we can look at migrating table content over, but that is some ways in the future.

Fixes #7834

Props needle, r-a-y

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12173 cdf35c40-ae34-48e0-9cc9-0c9da1808c22