Navigation: Ensure subnav menus can be registered before the parent nav is registered.
In #7659, a bug was fixed where restricted subnav menus should be
redirected instead of throwing a 404.  The side-effect of the fix is this
broke how plugins can register subnav menus before the parent nav is
created.

This commit rectifies the problem by not bailing out if the parent nav
cannot be found during subnav menu registration.  Includes a unit-test.

Props dontdream, boonebgorges, r-a-y.

Fixes #7931 (3.x branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12194 cdf35c40-ae34-48e0-9cc9-0c9da1808c22