Separate action logic from permission logic in bp-legacy joinleave_group AJAX callback.
Prior to [11776], join/leave group button intent was calculated based
directly on the user's group membership status. After that changeset,
the intent was calculated based on permission checks, which encompassed
the membership checks but also some additional information. This caused
problems when a user - such as a super admin - had *permission* to
join a group, but was not actually *trying* to join a group (but instead
had clicked Request Membership).

Merges [12198] to the 3.0 branch.

See #7936.


git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12199 cdf35c40-ae34-48e0-9cc9-0c9da1808c22