Render the Message compose view on user's profile header button click
In BP Nouveau, we use Backbone views to display the BP Messages component screens. When the user's profile header Private Message button is clicked, we need to make sure to populate the Send To input field with the corresponding username.
This commit also fix a regression introduced in r12115 where I forgot to update this specific button link.

Props RT77, r-a-y, DJPaul

See #7908 (Branch 3.0)

git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12249 cdf35c40-ae34-48e0-9cc9-0c9da1808c22