Nouveau: Change AJAX template loading event emission timing.
In Nouveau, a useful event, `bp_ajax_request`, is emitted when a template
is updated wth new content. However, the new content is eased in using
a fadeIn transition. This changeset moves the event to the fadeIn complete
so that any actions tied to `bp_ajax_request` will be operating on
visible and full-size elements.

Fixes #8128.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12447 cdf35c40-ae34-48e0-9cc9-0c9da1808c22