Make sure users can register when the xProfile component is not active
In version 5.0.0, we've introduced an improvement to use the display name as the salutation in signup emails (see [12349]). To do so we rely on an xProfile's component function. When this component is not active, the function is also not available causing a fatal error and preventing the activation email to be sent. To fix this issue we are now checking the component is active before using the function to return the display name.

See #8174 (Trunk)


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12503 cdf35c40-ae34-48e0-9cc9-0c9da1808c22