Admin: Make external links in BP Settings > Page accessible.
The `View` link beside each new activated component opens up in a new tab to 
ensure that a user may view the new BP page generated in the frontend without 
losing the selection/s in the settings page before the user has saved such.

This fix provides hidden text for users of screen readers and an icon for 
sighted users to indicate that the link will open in a new tab. 

Props audrasjb, imath.
Fixes #8235.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12550 cdf35c40-ae34-48e0-9cc9-0c9da1808c22