Don't delete user data on `delete_user` hook on Multisite.
The new function `bp_remove_user_data_on_delete_user_hook()`, which defaults
to `false` on Multisite and `true` otherwise, helps BuddyPress to differentiate
between the use of `wp_delete_user()` to delete a user account from an
installation (typical on non-Multisite) and its use to remove a user from
a site in a Multisite instance.

Props imath.

Fixes #8175.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12605 cdf35c40-ae34-48e0-9cc9-0c9da1808c22