Settings: Do not show "View" links for missing pages.
This commit adds `get_post()` checks to an existing comparison to ensure that some UI elements are not incorrectly shown. This bug could arise if component pages are trashed & deleted after BuddyPress is deactivated, causing a mismatch between the `bp-pages` option and the pages that exist on the site.

Fixes #8362.



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12718 cdf35c40-ae34-48e0-9cc9-0c9da1808c22